Functions have been moved to functions.py and module docstrings added
